+++ config/params.php
@@ -4,11 +4,18 @@ declare(strict_types=1);
 
 use Botasis\Runtime\Console\GetUpdatesCommand;
 use Botasis\Runtime\Console\SetTelegramWebhookCommand;
+use Botasis\Runtime\Handler\DummyUpdateHandler;
 
 return [
-    'botasis/telegram-bot' => [
+    'botasis/runtime' => [
         'bot token' => '',
         'errors to ignore' => [],
+        'fallback handler' => DummyUpdateHandler::class,
+        'request tags' => [
+            'success' => [],
+            'error' => [],
+        ],
+        'routes' => [],
     ],
     'yiisoft/yii-console' => [
         'commands' => [