+++ src/Response/Response.php
@@ -6,11 +6,12 @@ namespace Botasis\Runtime\Response;
 
 use Botasis\Client\Telegram\Request\CallbackResponse;
 use Botasis\Client\Telegram\Request\TelegramRequestInterface;
+use Botasis\Runtime\Request\TelegramRequestEnriched;
 use Botasis\Runtime\Update\Update;
 
 final class Response implements ResponseInterface
 {
-    /** @var TelegramRequestInterface[] */
+    /** @var TelegramRequestEnriched[] */
     private array $requests = [];
 
     private bool $hasCallbackResponse = false;
@@ -27,37 +28,38 @@ final class Response implements ResponseInterface
         return $instance;
     }
 
-    public function withRequest(TelegramRequestInterface $request): ResponseInterface
+    public function withRequest(TelegramRequestInterface $request, string ...$tags): ResponseInterface
     {
         $instance = clone $this;
+        $enriched = new TelegramRequestEnriched($request, ...$tags);
 
         if ($request->getMethod() === CallbackResponse::METHOD) {
             $instance->hasCallbackResponse = true;
-            array_unshift($instance->requests, $request);
+            array_unshift($instance->requests, $enriched);
         } else {
-            $instance->requests[] = $request;
+            $instance->requests[] = $enriched;
         }
 
         return $instance;
     }
 
-    public function withRequestReplaced(TelegramRequestInterface $search, ?TelegramRequestInterface $replace): ResponseInterface
+    public function withRequestReplaced(TelegramRequestInterface $search, ?TelegramRequestInterface $replace, string ...$replaceTags): ResponseInterface
     {
         $instance = clone $this;
 
         $requests = $this->requests;
         foreach ($requests as $index => $request) {
-            if ($request === $search) {
+            if ($request->request === $search) {
                 if ($replace === null) {
                     unset($requests[$index]);
-                } elseif ($request->getMethod() !== CallbackResponse::METHOD && $replace->getMethod() === CallbackResponse::METHOD) {
+                } elseif ($request->request->getMethod() !== CallbackResponse::METHOD && $replace->getMethod() === CallbackResponse::METHOD) {
                     unset($requests[$index]);
-                    array_unshift($requests, $replace);
-                } elseif($replace->getMethod() !== CallbackResponse::METHOD && $request->getMethod() === CallbackResponse::METHOD) {
+                    array_unshift($requests, new TelegramRequestEnriched($replace, ...$replaceTags));
+                } elseif($replace->getMethod() !== CallbackResponse::METHOD && $request->request->getMethod() === CallbackResponse::METHOD) {
                     unset($requests[$index]);
-                    $requests[] = $replace;
+                    $requests[] = new TelegramRequestEnriched($replace, ...$replaceTags);
                 } else {
-                    $requests[$index] = $replace;
+                    $requests[$index] = new TelegramRequestEnriched($replace, ...$replaceTags);
                 }
 
                 break;