+++ tests/Middleware/MiddlewareDispatcherTest.php
@@ -11,6 +11,7 @@ use Botasis\Runtime\Middleware\Event\AfterMiddleware;
 use Botasis\Runtime\Middleware\Event\BeforeMiddleware;
 use Botasis\Runtime\Middleware\MiddlewareDispatcher;
 use Botasis\Runtime\Middleware\MiddlewareFactory;
+use Botasis\Runtime\Request\TelegramRequestEnriched;
 use Botasis\Runtime\Response\Response;
 use Botasis\Runtime\Response\ResponseInterface;
 use Botasis\Runtime\Tests\Middleware\Support\FailMiddleware;
@@ -39,7 +40,7 @@ final class MiddlewareDispatcherTest extends TestCase
         );
 
         $response = $dispatcher->dispatch($update, $this->getRequestHandler());
-        $this->assertSame('middleware-id', $response->getRequests()[0]?->id);
+        $this->assertSame('middleware-id', $response->getRequests()[0]?->request->id);
     }
 
     public function testArrayMiddlewareCall(): void
@@ -53,7 +54,7 @@ final class MiddlewareDispatcherTest extends TestCase
         $dispatcher = $this->createDispatcher($container)->withMiddlewares([TestController::class, 'index']);
 
         $response = $dispatcher->dispatch($update, $this->getRequestHandler());
-        $this->assertSame('test message', $response->getRequests()[0]?->text);
+        $this->assertSame('test message', $response->getRequests()[0]?->request->text);
     }
 
     public function testMiddlewareFullStackCalled(): void
@@ -74,7 +75,7 @@ final class MiddlewareDispatcherTest extends TestCase
         $dispatcher = $this->createDispatcher()->withMiddlewares($middleware1, $middleware2);
 
         $response = $dispatcher->dispatch($update, $this->getRequestHandler());
-        $this->assertSame('middleware1', $response->getRequests()[0]?->id);
+        $this->assertSame('middleware1', $response->getRequests()[0]?->request->id);
     }
 
     public function testMiddlewareStackInterrupted(): void
@@ -95,7 +96,7 @@ final class MiddlewareDispatcherTest extends TestCase
         $dispatcher = $this->createDispatcher()->withMiddlewares($middleware1, $middleware2);
 
         $response = $dispatcher->dispatch($update, $this->getRequestHandler());
-        $this->assertSame('first', $response->getRequests()[0]?->id);
+        $this->assertSame('first', $response->getRequests()[0]?->request->id);
     }
 
     public function testEventsAreDispatched(): void
@@ -190,7 +191,7 @@ final class MiddlewareDispatcherTest extends TestCase
         $dispatcher = $dispatcher->withMiddlewares(TestMiddleware::class);
         $response = $dispatcher->dispatch($update, $this->getRequestHandler());
 
-        self::assertSame('42', $response->getRequests()[0]?->id);
+        self::assertSame('42', $response->getRequests()[0]?->request->id);
     }
 
     private function getRequestHandler(): UpdateHandlerInterface