+++ tests/Middleware/Support/UseParamsController.php
@@ -6,6 +6,7 @@ namespace Botasis\Runtime\Tests\Middleware\Support;
 
 use Botasis\Client\Telegram\Request\Message\Message;
 use Botasis\Client\Telegram\Request\Message\MessageFormat;
+use Botasis\Runtime\Request\TelegramRequestEnriched;
 use Botasis\Runtime\Response\ResponseInterface;
 use Botasis\Runtime\Update\Update;
 use Botasis\Runtime\UpdateHandlerInterface;
@@ -16,8 +17,6 @@ final class UseParamsController
     {
         return $handler
             ->handle($request)
-            ->withRequest(
-                new Message('message text', MessageFormat::TEXT, $request->chat->id)
-            );
+            ->withRequest(new Message('message text', MessageFormat::TEXT, $request->chat->id));
     }
 }